Prints an integer

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -6,7 +6,7 @@
  */
 void print_number(int n)
 {
-	int rev = 0, rem = 0, neg = 0;
+	long rev = 0, rem = 0, neg = 0;
 
 	if (n > 0)
 	{
